requirements.txt
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
chromadb==0.4.15
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sentence-transformers==2.2.2
gradio==4.19.2
ollama==0.1.6
torch==2.2.0
transformers==4.37.2
huggingface-hub==0.19.4
pandas==2.2.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.9